+++ src/features/discussions/reply/ui/Replies.tsx
@@ -3,7 +3,7 @@ import { ProblemId } from '@/shared';
 import { Spinner } from '@/shared/ui/loading-indicators';
 import Reply from './Reply';
 import { useRepliesQuery } from '@/entities/discussionReplies';
-import ReplyForm from '../../ReplyForm';
+import ReplyForm from './ReplyForm';
 
 interface RepliesProps {
   problemId: ProblemId;
