+++ src/features/discussions/vote/index.ts
@@ -1,0 +1,3 @@
+/**vote */
+
+export { default as Vote } from './ui/Vote';
