+++ src/features/discussions/vote/ui/Vote.tsx
